iconhandler: Get theme contents from the .deb file, not Contents
Sometimes Contents are updated only very rarely, for example on
Launchpad they are only done once per week. The .deb file is always up
to date - for icon themes use the .deb file. We still need index.theme
to be in the Contents file to know in *which* deb file to look.